+++ e2e-tests/test-applications/angular-19-test-app/src/app/app.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+
+@Component({
+  selector: 'app-root',
+  imports: [RouterOutlet],
+  templateUrl: './app.component.html',
+})
+export class AppComponent {
+  title = 'angular-19-test-app';
+}